requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
Jinja2==3.1.3
jiwer==3.0.4
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
pyreadline3==3.4.1
pytools==2024.1.11
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
seaborn==0.13.2
sentry-sdk==2.11.0
setproctitle==1.3.3
setuptools==71.0.4
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.43.3
tzdata==2024.1
ultralytics-thop==2.0.0
ultralytics==8.1.34
urllib3==2.2.2
wandb==0.17.5
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
fastapi==0.111.0
uvicorn==0.30.1
git+https://github.com/THU-MIG/yolov10.git
easyocr==1.7.1